requirements.txt
streamlit
PyQt6
pandas
pyautogui
pyscreeze
pygetwindow
mouseinfo
pyrect
pyperclip
pillow
opencv-python
easyocr
google-generativeai
python-dotenv
numpy
watchdog
torch
torchvision
scikit-image
ninja
pyclipper
Shapely
keyboard
